requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
PyYAML==6.0.1
lxml==4.9.3
titlecase==2.4.1
